refactor: move ConnectionManager to utils
